requirements.txt
django
pillow
django-crispy-forms
django-rest-framework